feat: Add docstrings
chore: lints
